% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{check_cols}
\alias{check_cols}
\title{Check whether expected columns are present}
\usage{
check_cols(df, df_name, df_type)
}
\arguments{
\item{df}{Data frame. Data frame to check.}

\item{df_name}{Character. Name of data frame as a character.}

\item{df_type}{Character. One of: "unitesFixes", "releve", "unitesautre1", "autre1releve", "unitesautre2", "autre2releve"}
}
\description{
Check whether expected columns are present
}
