% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{combine_nsu_data}
\alias{combine_nsu_data}
\title{Fusionner l'ensemble des bases NSU}
\usage{
combine_nsu_data(dir_in, dir_regexp = "_STATA_", data_type, dir_out)
}
\arguments{
\item{dir_in}{Character. Root directory where food group sub-directories are located.}

\item{dir_regexp}{Character. Regular expression to identify folders over which to iterate.}

\item{data_type}{Character. One of: "consumption", "production".
For data from markets, "consumption"; for data from farmers, "production".}

\item{dir_out}{Character. Directory where combined files should be saved.}
}
\description{
Créer les bases suivantes sous format Stata:
}
\details{
\itemize{
\item unitesFixes
\item releve
\item unitesautre1
\item autre1releve
\item unitesautre2
\item autre2releve
}

Ces bases seront créées en:
\itemize{
\item Parcourant les répertoires qui correspondent à la chaîne regex \code{dir_regexp}.
\item Fusionnant les bases cibles
\item "Fusionnant" les étiquettes de valeurs pour les variables étiquettées
\item Sauvegardant les bases résultantes dans le répertoire \code{dir_out}
}
}
