% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{combine_nsu_in_dir}
\alias{combine_nsu_in_dir}
\title{Combine NSU all data of same type in the same directory}
\usage{
combine_nsu_in_dir(dir_in, df_type)
}
\arguments{
\item{dir_in}{Character. Directory path whose files should be combined.}

\item{df_type}{Character. One of: "unitesFixes", "releve", "unitesautre1", "autre1releve", "unitesautre2", "autre2releve"}
}
\description{
Combine NSU all data of same type in the same directory
}
