% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{correct_col_type}
\alias{correct_col_type}
\title{Change column to expected type}
\usage{
correct_col_type(df, df_type)
}
\arguments{
\item{df}{Data frame whose columns types to check and correct.}

\item{df_type}{Character. One of: "unitesFixes", "releve", "unitesautre1", "autre1releve", "unitesautre2", "autre2releve"}
}
\value{
Data frame with column types corrected where applicable.
}
\description{
First, check whether the column is of the expected type
Then, if not, change the column's type
}
