% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{create_product_dirs}
\alias{create_product_dirs}
\title{Créer les répertoires pour les produits avec images}
\usage{
create_product_dirs(inventory_df, dir)
}
\arguments{
\item{inventory_df}{Data frame. Contains an inventory of the unique product-units with images.}

\item{dir}{Character. Directory where product directories should be created.}
}
\description{
Créer un répertoire pour chaque produit avec image.
}
\details{
If the folder does not exist, create it. If the folder already exists, move to the next folder.
}
