% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{create_strata_dirs}
\alias{create_strata_dirs}
\title{Create dirs for strata}
\usage{
create_strata_dirs(inventory_df, market_path, strata_vars, dir)
}
\arguments{
\item{inventory_df}{Data frame. Contains an inventory of the unique product-units with images.}

\item{market_path}{Character. File path to the combined market file.}

\item{strata_vars}{Character vector. Names of the variables that together form strata.}

\item{dir}{Character. Directory where unit directories should be created.}
}
\description{
Create directories for each strata within each product-unit directory
}
\details{
Strata are the user-defined set of variables specified in \code{strata_vars}.
The directory names will be hyphen-separated combinations of the variables that compose a strata (e.g., \code{"{region} - {urban_rural}"})
}
