% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{create_unit_dirs}
\alias{create_unit_dirs}
\title{Create unit directories}
\usage{
create_unit_dirs(inventory_df, dir)
}
\arguments{
\item{inventory_df}{Data frame. Contains an inventory of the unique product-units with images.}

\item{dir}{Character. Directory where unit directories should be created.}
}
\description{
Create one directory for each product-unit with an image.
}
\details{
If the folder does not exist, create it. If the folder already exists, move to the next folder.
}
