% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{extract_labels}
\alias{extract_labels}
\title{Extract labels from all labelled variables}
\usage{
extract_labels(df)
}
\arguments{
\item{df}{Data frame whose value labels to extract}
}
\value{
List of value labels. Each element is a named numeric vector. Each vector consists of a named values. The names correspond to character labels and the values to the numerical values.

Side-effect: create list object in global environment containing labels for all variables
}
\description{
First, determines which variables are labelled.
Then, extracts the labels
Next, creates a named list: names are variable names; values are associated labels.
Then, creates a label object in the global environment whose name is the \verb{\{df\}_lbls}.
}
