% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{find_market_file}
\alias{find_market_file}
\title{Find the market file for the food group}
\usage{
find_market_file(
  dir,
 
    pattern_to_exclude = "^interview__|^assignment__|^autre2releve_|^unites[Aa]utre2|^autre1releve_|^unites[Aa]utre1|^releve_|^unitesFixes_"
)
}
\arguments{
\item{dir}{Character. Directory whose files to search for the market file.}

\item{pattern_to_exclude}{Character. Regular expression of files that are not the market file.}
}
\value{
Character. Name of the market file, with extension
}
\description{
First, filter files:
\itemize{
\item those with a .dta extension
\item those that do not start with prefixes associated with other files (e.g., "interview__", "assignment__", "unitesFixes", etc.)
}
}
\details{
Then, retain as the market file the only file remaining.
}
