% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{get_interview_ids}
\alias{get_interview_ids}
\title{Get interview IDs}
\usage{
get_interview_ids(path)
}
\arguments{
\item{path}{Character. Full path to combined data files of interest
(i.e., unitesFixes.dta, unitesautre1.dta, and unitesautre2.dta)}
}
\description{
Get interview ID variables for all files with images.
This data will the image sorting functions to construct a path to the strata folders.
}
