% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_matching_files}
\alias{get_matching_files}
\title{Get vector of file names matching a pattern}
\usage{
get_matching_files(dir, pattern)
}
\arguments{
\item{dir}{Character vector. File path to directory}

\item{pattern}{Character vector. Regular expression describing files to return}
}
\value{
Character vector. File names with \code{.dta} extension
}
\description{
Returns a character vector of file names that match a regex pattern in the target directory
}
