% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{inventory_product_units}
\alias{inventory_product_units}
\title{Créer un inventaire des produits-unités avec images}
\usage{
inventory_product_units(
  dir,
  unitesFixes = "unitesFixes.dta",
  unitesautre1 = "unitesautre1.dta",
  unitesautre2 = "unitesautre2.dta"
)
}
\arguments{
\item{dir}{Character vector. Chemin du répertoire où se trouvent les bases de données resorties de la fusion des bases.}

\item{unitesFixes}{Character vector. Nom du fichier des unités fixes.}

\item{unitesautre1}{Character vector. Nom du fichier des unités connues mais associées avec d'autres produits.}

\item{unitesautre2}{Character vector. Nom du fichier des unités inconnues d'avance.}
}
\value{
Data frame des produits-unités distincts avec images
}
\description{
Puiser dans les bases NSU et identifier les produits-unités distincts et avec images.
}
\details{
Les produits-unités avec images sont définis comme ceux avec un texte non-vide au niveau des questions q109, q114_autre1, ou q119_autre2
}
