% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{label_variables}
\alias{label_variables}
\title{Label variables}
\usage{
label_variables(df, labels_list)
}
\arguments{
\item{df}{Data frame}

\item{labels_list}{List that contains the consolidated labels}
}
\value{
Data frame with labelled columns
}
\description{
Apply labels to all variables in the data frame that need them. To do this:
}
\details{
First, determine the variables that have labels for them in the list of labels.
Then, apply those labels to each column
}
