% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{rename_variables}
\alias{rename_variables}
\title{Rename variables to harmonized names}
\usage{
rename_variables(df, df_type)
}
\arguments{
\item{df}{Data frame whose columns to rename.}

\item{df_type}{Character. One of: "unitesFixes", "releve", "unitesautre1", "autre1releve", "unitesautre2", "autre2releve"}
}
\value{
Data frame with whose columns have harmonized names.
}
\description{
First, changes variable names by removing the product-specific component, thereby leaving harmonized names.
Then, overwrite the input df in the global environment with an updated df.
}
