% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{save_nsu_data}
\alias{save_nsu_data}
\title{Save data to disk}
\usage{
save_nsu_data(df, df_type, dir)
}
\arguments{
\item{df}{Data frame to save to disk.}

\item{df_type}{Character. One of: "unitesFixes", "releve", "unitesautre1", "autre1releve", "unitesautre2", "autre2releve"}

\item{dir}{Character. Directory where to save data.}
}
\description{
Save data to disk in Stata format
}
