% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{sort_images_obs}
\alias{sort_images_obs}
\title{Sort images for an interview}
\usage{
sort_images_obs(dir_in, unit_labels, dir_out)
}
\arguments{
\item{dir_in}{Character. Folder for a single interview whose images to sort}

\item{unit_labels}{Named numeric vector.}

\item{dir_out}{Character. Root folder of destination folder for sorted images.}
}
\description{
This workhorse function for \code{sort_images()} works within an interview folder of the exported images and copies images to their destination directory based on the product-unit found in the file name.
}
\details{
First, it determines where files should go by:
\itemize{
\item Creating a list of files in the folder
\item Parsing the file name into product name and unit code components
\item Constructs a file name that includes the file's \code{interview__key}
\item Constructing a destination file path
}

Then, it moves the file to its destination directory by:
\itemize{
\item Checking whether the directory exists
\item If not, issuing a warning
\item If so, copying the file
}
}
