% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{sort_images_product}
\alias{sort_images_product}
\title{Sort images for a product group}
\usage{
sort_images_product(dir_in, unit_labels, dir_out)
}
\arguments{
\item{dir_in}{Character. Folder for a single product group whose images to sort}

\item{unit_labels}{Named numeric vector.}

\item{dir_out}{Character. Root folder of destination folder for sorted images.}
}
\description{
Applies \code{sort_images_obs()} over a list of product group folders
}
