% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{create_image_dirs}
\alias{create_image_dirs}
\title{Créer un répertoire pour chaque produit-unité}
\usage{
create_image_dirs(inventory_df, dir)
}
\arguments{
\item{inventory_df}{Data frame. Base créée par \code{inventory_product_units()} qui fait l'inventaire complet des produits-unités des bases de l'enquête NSU.}

\item{dir}{Character. Répertoire racine où les répertoires produit-unité doivent être créés.}
}
\description{
Créer un répertoire pour chaque produit-unité
}
